+++ packages/types/src/index.ts
@@ -53,6 +53,7 @@ export type PackageJSON = {
     access?: AccessType;
     directory?: string;
     registry?: string;
+    [registry: `${string}:registry`]: string;
   };
 };
 
@@ -75,6 +76,8 @@ export type Config = {
   baseBranch: string;
   githubRelease?: boolean;
   changedFilePatterns: readonly string[];
+  /** When false, Changesets won't format with Prettier */
+  prettier: boolean;
   /** Features enabled for Private packages */
   privatePackages: PrivatePackages;
   /** The minimum bump type to trigger automatic update of internal dependencies that are part of the same release */
@@ -101,6 +104,7 @@ export type WrittenConfig = {
   baseBranch?: string;
   changedFilePatterns?: readonly string[];
   githubRelease?: boolean;
+  prettier?: boolean;
   /** Opt in to tracking non-npm / private packages */
   privatePackages?:
     | false